+++ packages/eslint-plugin/src/utils/index.ts
@@ -2,5 +2,4 @@ export { createNoExportedRule } from './create-no-exported-rule';
 export { checkIfCompiledExport as validateDefinition } from './create-no-exported-rule/check-if-compiled-export';
 export { createNoTaggedTemplateExpressionRule } from './create-no-tagged-template-expression-rule';
 export { CssMapObjectChecker, getCssMapObject } from './css-map';
-export { isCss, isCssMap, isKeyframes, isCxFunction } from './is-compiled-import';
-export { isStyled } from '../rules/no-styled-tagged-template-expression/utils';
+export { isCss, isCssMap, isKeyframes, isCxFunction, isStyled } from './is-compiled-import';
